backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.10.1

# Speech & Voice Processing
openai-whisper==20250625
speechbrain==1.0.0
librosa==0.10.1
soundfile==0.12.1
numpy==1.26.0

# Real-time & WebSocket
websockets==12.0

# HTTP & API Clients
httpx==0.25.2
aiohttp==3.9.1

# Caching & Queue
redis==5.0.1

# Utilities
pydantic-extra-types==2.1.0
python-slugify==8.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.2
httpx==0.25.2

# Development
black==23.12.1
flake8==6.1.0
isort==5.13.2
mypy==1.7.1